pyproject.toml
[tool.poetry]
name = "ml-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
numpy = "^1.25.1"
pandas = "^2.0.3"
lightgbm = "^4.0.0"
tensorflow-io-gcs-filesystem = "0.27.0"
tensorflow = ">=2.10,<2.11"
keras = ">=2.10,<2.11"
scipy = "^1.11.1"
matplotlib = "^3.7.2"
scikit-learn = "^1.3.0"
hydra-core = "^1.3.2"
catboost = "^1.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
